+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 -   Added `navigationLeaves` option to remove branches from the navigation tree, #2382.
+-   Added `sortEntryPoints` option (defaults to true) to allow disabling entry point sorting, #2393.
 -   Improved support for multi-word searches, #2400.
 
 ### Bug Fixes